Add files via upload
This is a small shell script that I wrote up (because I have never done one before) to wipe docker like we do in step 0.  I have also included the command "docker-compose image prune -a && yes".  

I kept looking up the command and thought it would be easier to have a shell script instead of always looking it up.